Fix `res.json()` Returning String Instead of Object as of Next.js 12.1.6.
Before the update to Next.js 12.1.6, JSON was emitted unformatted by Next.js's
`res.json()`.

Therefore, to fix this problem, the static method `JSON.prettyPrint()` was
implemented, which adds whitespace to the JSON response to improve readability
for users.

However, in a Next.js bug report, Vincent Voyer reported, that when he was
using

  const result = "yes";
  res.json(result);

the response body was `yes` instead of `"yes"` [1]. It has been fixed by Pull
Request [2].

Since the `JSON.prettyPrint()` method, as described before, returns a whitespace
formatted string instead of an object, the emitted JSON response is affected by
this change. Therefore, the data to be emitted must no longer be formatted by
`JSON.prettyPrint()`.

[1] https://github.com/vercel/next.js/issues/21749
[2] https://github.com/vercel/next.js/pull/36186

Signed-off-by: Stefan Kühnel <git@stefankuehnel.com>